Update project.json
Would be great to give credit to those who contributed to this project by running AraCred on the repos we've forked and modified. Added TheSource to see how that affects Cred scores. If this throws things out of balance, however, then we might need to wait to integrate it otherwise new contributors won't get any meaningful amounts of Cred for a long time lol